+++ example/example.py
@@ -1,28 +1,69 @@
 #!/usr/bin/env python
 
 import sys
-sys.path.append("../")
+sys.path.insert(0,"../")
 
 import logging
 logging.basicConfig(level=logging.DEBUG)
 
+import os, socket
+import win32net  # install package pypiwin32 
 from flask import Flask
-from flask import render_template
-from flask_sspi import init_sspi
-from flask_sspi import requires_authentication
+from flask import render_template, session, g, url_for
+from flask_sspi import *
 
 DEBUG=True
 
 app = Flask(__name__)
 app.secret_key = 'efca0226-1746-43f6-92ac-1975e1eea085'
 
+init_sspi(app)
 
 @app.route("/")
-@requires_authentication
-def index(user):
-    print("index")
-    return render_template('index.html', user=user)
+def index():
+    links = []
+    for rule in app.url_map.iter_rules():
+        # Filter out rules we can't navigate to in a browser
+        # and rules that require parameters
+        if "GET" in rule.methods and has_no_empty_params(rule):
+            url = url_for(rule.endpoint, **(rule.defaults or {}))
+            if rule.endpoint.startswith("test_") or rule.endpoint.startswith("Blue"):
+                links.append((url, rule.endpoint))
+    # links is now a list of url, endpoint tuples
+    return render_template("index.html", links=links)
 
+@app.route("/requires_authentication")
+@requires_authentication
+def test_requires_authentication(user):
+    print("user: " + user)
+    return render_template('test.html', current_user=user, function = test_requires_authentication.__name__)
+
+@app.route("/authenticate")
+@authenticate
+def test_authenticate():
+    print("user: " + g.current_user)
+    return render_template('test.html', function = test_authenticate.__name__) # current_user is also defined in Jinja's context
+
+@app.route("/impersonate")
+@authenticate
+def test_impersonate():
+    with Impersonate(): # Some functions like print do not work with this. Hard to debug.
+        user = os.getlogin()
+    domain = '.'.join(socket.getfqdn().split('.')[1:])
+    groups = win32net.NetUserGetGroups(domain, user)
+    groups = ', '.join(map(lambda r:r[0], groups))
+    print("user: " + os.getlogin())
+    return render_template('test_groups.html', current_user=user, groups=groups, function = test_impersonate.__name__) # current_user is also defined in Jinja's context
+
+def has_no_empty_params(rule):
+    defaults = rule.defaults if rule.defaults is not None else ()
+    arguments = rule.arguments if rule.arguments is not None else ()
+    return len(defaults) >= len(arguments)
+
+########### Registrations ###########
+from blueprint1.main import Blueprint1
+########### Register apps ###########
+app.register_blueprint(Blueprint1, url_prefix='/blueprint1')
 
 if __name__ == '__main__':
     init_sspi(app)
